rename to lngsToLoad
