+++ src/utils/index.js
@@ -5,4 +5,4 @@ export { default as lngPathCorrector } from './lng-path-corrector'
 export { default as lngPathDetector } from './lng-path-detector'
 export { default as localeSubpathRequired } from './locale-subpath-required'
 export { default as redirectWithoutCache } from './redirect-without-cache'
-export { default as lngsToLoadDetector } from './lngs-to-load-detector'
+export { default as lngsToLoad } from './lngs-to-load'
